crud operations

+++ src/main/java/com/example/HumanResourceApplication/repository/RegionRepository.java
@@ -3,5 +3,5 @@ package com.example.HumanResourceApplication.repository;
 import com.example.HumanResourceApplication.entity.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegionRepository extends JpaRepository<Region  , Long> {
+public interface RegionRepository extends JpaRepository<Region  , Integer> {
 }
